Refactor: Todo lo que concierne productos fue movido a su propio package.
