+++ src/main/java/org/main/Producto/CocaCola.java
@@ -1,4 +1,4 @@
-package org.main;
+package org.main.Producto;
 
 public class CocaCola extends Bebida {
     public CocaCola(String serie) {
