+++ src/main/java/org/main/Producto/Dulce.java
@@ -1,4 +1,4 @@
-package org.main;
+package org.main.Producto;
 
 public abstract class Dulce extends Producto {
     public Dulce(String serie) {
