+++ src/main/java/org/main/Producto/Super8.java
@@ -1,4 +1,4 @@
-package org.main;
+package org.main.Producto;
 
 public class Super8 extends Dulce {
     public Super8(String serie) {
